pyproject.toml
[project]
name = "minute-bot"
version = "0.1.0"
description = "Agentic Meeting Memory System with Speaker-Aware Action Extraction"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "redis>=5.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "supabase>=2.0.0",
    "gunicorn>=21.0.0",
    "pyaudio>=0.2.14",
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "pyannote.audio>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
minute-bot = "minute_bot.api:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
